extract generation config into core microcrate
