+++ Cargo.toml
@@ -37,7 +37,8 @@ members = [
   "crates/bitnet-intel-gpu-id",  # SRP: Intel GPU/Arc identification heuristics
   "crates/bitnet-logits",        # SRP: pure logits transform functions
   "crates/bitnet-logits-filters", # SRP: reusable logits/probability filtering transforms
-  "crates/bitnet-generation",    # SRP: decode-loop stopping logic & generation events
+  "crates/bitnet-generation",    # SRP: decode-loop stopping logic + generation contract re-exports
+  "crates/bitnet-generation-config-core", # SRP: reusable generation request config contract
   "crates/bitnet-generation-events-core", # SRP: reusable generation stream event/stat contracts
   "crates/bitnet-generation-stop-core", # SRP: reusable stop criteria/check core
   "crates/bitnet-kv-cache-policy-core", # SRP: KV-cache sequence transition decision policy
@@ -165,6 +166,7 @@ default-members = [
   "crates/bitnet-probability",
   "crates/bitnet-logits-filters",
   "crates/bitnet-generation",
+  "crates/bitnet-generation-config-core",
   "crates/bitnet-generation-events-core",
   "crates/bitnet-generation-stop-core",
   "crates/bitnet-kv-cache-policy-core",
