+++ crates/bitnet-generation-config-core/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "bitnet-generation-config-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "Reusable generation request config contract for BitNet inference"
+rust-version.workspace = true
+
+[dependencies]
+bitnet-generation-stop-core = { path = "../bitnet-generation-stop-core", version = "0.2.1-dev" }
+serde.workspace = true
+
+[lints]
+workspace = true
